pyproject.toml
[project]
name = "ghostroute-security-homepage-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Ayobami6",email = "ayobamidele006@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.9,<0.116.0)",
    "sparky-utils (>=1.1.8,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "django (>=5.1.6,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
